feature: make app icon change with app theme
